Move sink/stream
