+++ src/raw/io/input/into_sink.rs
