+++ src/raw/io/input/into_stream.rs
