+++ src/raw/io/mod.rs
@@ -25,15 +25,6 @@ pub mod output;
 mod api;
 pub use api::BytesCursor;
 
-mod into_stream;
-pub use into_stream::into_stream;
-
-mod into_sink;
-pub use into_sink::into_sink;
-
-mod into_writer;
-pub use into_writer::into_writer;
-
 mod read_observer;
 pub use read_observer::observe_read;
 pub use read_observer::ReadEvent;
